improve PHPStan (#7)
* wip

* Fix styling

* pass level 5

* pass level 6

* pass level 7

* pass level 8

* pass level max

+++ src/Http/Livewire/DeletePost.php
@@ -2,22 +2,30 @@
 
 namespace Mito\Http\Livewire;
 
+use Illuminate\View\View;
 use LivewireUI\Modal\ModalComponent;
 use Mito\Models\Post;
 
 class DeletePost extends ModalComponent
 {
     public int|Post $post;
-    public $type;
+    public string $type;
 
-    public function mount(Post $post)
+    public function mount(Post $post): void
     {
         $this->post = $post;
         $this->type = $post->status;
     }
 
+    /**
+     * @return \Illuminate\Http\RedirectResponse|void
+     */
     public function delete()
     {
+        if (! ($this->post instanceof Post)) {
+            return;
+        }
+
         $this->post->delete();
 
         return redirect()->to(route('mito.posts.index', ['type' => $this->type]));
@@ -28,7 +36,7 @@ class DeletePost extends ModalComponent
         return 'md';
     }
 
-    public function render()
+    public function render(): View
     {
         return view('mito::livewire.confirm-delete-post');
     }
